+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/presencetracing/locations/server/qrcodepostertemplate/QrCodePosterTemplateServer.kt
@@ -1,4 +1,4 @@
-package de.rki.coronawarnapp.eventregistration.events.server.qrcodepostertemplate
+package de.rki.coronawarnapp.presencetracing.locations.server.qrcodepostertemplate
 
 import androidx.annotation.VisibleForTesting
 import com.google.protobuf.InvalidProtocolBufferException
