+++ Corona-Warn-App/src/test/java/de/rki/coronawarnapp/eventregistration/checkins/qrcode/TraceLocationUserInputToTraceLocationMapperTest.kt
@@ -1,7 +1,7 @@
 package de.rki.coronawarnapp.eventregistration.checkins.qrcode
 
-import de.rki.coronawarnapp.eventregistration.events.TraceLocationUserInput
-import de.rki.coronawarnapp.eventregistration.events.toTraceLocation
+import de.rki.coronawarnapp.presencetracing.locations.TraceLocationUserInput
+import de.rki.coronawarnapp.presencetracing.locations.toTraceLocation
 import de.rki.coronawarnapp.presencetracing.checkins.qrcode.TraceLocation
 import de.rki.coronawarnapp.server.protocols.internal.pt.TraceLocationOuterClass
 import io.kotest.matchers.shouldBe
